+++ src/Interfaces/ReadCacheInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Sip\ReaderManager\Interfaces;
+
+interface ReadCacheInterface
+{
+    public function __construct(string $indexName, array $options);
+}