+++ src/Interfaces/ReaderFilterStorageInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Sip\ReaderManager\Interfaces;
+
+interface ReaderFilterStorageInterface
+{
+    public function __construct(string $name);
+
+    public function getStorageDirectory(): string;
+}